+++ mapmatching/geo/ops/linear_referencing.py
@@ -1,7 +1,8 @@
+import numba
 import numpy as np
 from shapely import LineString, Point
-
 from .distance import coords_seq_distance
+from .to_array import geoseries_to_coords, points_geoseries_2_ndarray
 
 def _check(point, line):
     res = linear_referencing(point, polyline, cut=False)
@@ -31,24 +32,26 @@ def plot(point, polyline, res):
 
     return ax
 
-def closest_point_on_segments(point:np.array, lines:np.array, eps=1e-9):
-    """_summary_
+def closest_point_on_segments(point:np.ndarray, lines:np.ndarray, eps=1e-9):
+    """Calculate the closest point p' and its params on each segments on a polyline.
 
     Args:
-        point (np.array): _description_
-        lines (np.array): _description_
-        eps (float, optional): _description_. Defaults to 1e-9.
+        point (np.array): Point (shape: [2,]). 
+        lines (np.array): Polyline in the form of coords sequence(shape: [n, 2]).
+        eps (float, optional): Defaults to 1e-9.
 
     Returns:
-        tuple: (proj, dist, ratio)
+        (array, array, array): proj, dist, ratio
     """
-    pq = lines[:, 1] - lines[:, 0]
+    segs = np.hstack([lines[:-1][:, np.newaxis],
+                      lines[1:][:, np.newaxis]])
+    pq = segs[:, 1] - segs[:, 0]
     d = np.power(pq, 2).sum(axis=1)
     d[d == 0] = eps
 
     x, y = point
-    dx = x - lines[:, 0, 0]
-    dy = y - lines[:, 0, 1]
+    dx = x - segs[:, 0, 0]
+    dy = y - segs[:, 0, 1]
     t = pq[:, 0] * dx + pq[:, 1] * dy
     
     ratio = t / d
@@ -56,45 +59,43 @@ def closest_point_on_segments(point:np.array, lines:np.array, eps=1e-9):
     ratio[ratio > 1] = 1
 
     offset = pq * ratio[:, np.newaxis]
-    proj =  offset + lines[:, 0]
+    proj =  offset + segs[:, 0]
     dist = np.linalg.norm(point - proj, axis=1)
 
     return proj, dist, ratio
 
+# @numba.jit
+def cut_lines(idx, proj, ratio, coords):
+    NONE_COORD = None
+    if idx == 0 and ratio == 0:
+        return NONE_COORD, coords
+    if idx == coords.shape[0] - 2 and ratio == 1:
+        return coords, NONE_COORD
+
+    if ratio == 0:
+        seg_0 = coords[:idx + 1] 
+        seg_1 = coords[idx:]    
+    elif ratio < 1:
+        seg_0 = np.concatenate([coords[:idx+1], [proj]]) 
+        seg_1 = np.concatenate([[proj], coords[idx+1:]])     
+    else:
+        seg_0 = coords[:idx+2]
+        seg_1 = coords[idx+1:]
+    
+    return seg_0, seg_1
+
 def linear_referencing(point:Point, polyline:LineString, cut=True, to_geom=False):
     # TODO vectorized
     # iterating through each segment in the polyline and returning the one with minimum distance
-    NONE_COORD = None
 
-    def cut_lines(idx, proj, ratio, coords):
-        if ratio == 0:
-            if idx == 0:
-                seg_0, seg_1 = NONE_COORD, coords
-            else:
-                seg_0 = coords[:idx + 1] 
-                seg_1 = coords[idx:]    
-        elif ratio < 1:
-            seg_0 = np.concatenate([coords[:idx+1], [proj]]) 
-            seg_1 = np.concatenate([[proj], coords[idx+1:]])     
-        else:
-            if idx == len(dists) - 1:
-                seg_0, seg_1 = coords, NONE_COORD
-            else:
-                seg_0 = coords[:idx+2]
-                seg_1 = coords[idx+1:]
-        
-        return seg_0, seg_1
-        
     p_coords = np.array(point.coords[0])
-    coords = np.array(polyline.coords)
-    l_coords = np.hstack([coords[:-1][:, np.newaxis],
-                          coords[1:][:, np.newaxis]])
+    l_coords = np.array(polyline.coords)
 
     projs, dists, ratios = closest_point_on_segments(p_coords, l_coords)
     idx = np.argmin(dists)
     proj = projs[idx]
     ratio = ratios[idx]
-    len_np, total_len = coords_seq_distance(coords)
+    len_np, total_len = coords_seq_distance(l_coords)
     offset = len_np[:idx].sum() + len_np[idx] * ratio
 
     res = {}
@@ -103,11 +104,117 @@ def linear_referencing(point:Point, polyline:LineString, cut=True, to_geom=False
     if not cut:
         res['offset'] = offset
     else:
-        seg_0, seg_1 = cut_lines(idx, proj, ratio, coords)
-        res['seg_0'] = LineString(seg_0) if to_geom else seg_0
-        res['seg_1'] = LineString(seg_1) if to_geom else seg_1
+        seg_0, seg_1 = cut_lines(idx, proj, ratio, l_coords)
+        if to_geom:
+            seg_0 = LineString(seg_0)
+            seg_1 = LineString(seg_1)
+        res['seg_0'] = seg_0
+        res['seg_1'] = seg_1
         res['len_0'] = offset
         res['len_1'] = total_len - offset
+
+    return res
+
+# @numba.jit
+def lines_to_matrix(lines, n_rows, n_cols):
+    _lines = np.zeros((n_rows, n_cols, 2))
+    mask = np.ones((n_rows, n_cols), dtype=np.bool_)
+
+    for i, line in enumerate(lines):
+        n = len(line)
+        _lines[i, :n] = line
+        _lines[i, n:] = line[-1]
+        mask[i, n:] = 0
+    
+    return _lines, mask
+
+# @numba.jit
+def cut_line(idx, proj, ratio, coords):
+    NONE_COORD = None
+    if idx == 0 and ratio == 0:
+        return NONE_COORD, coords
+    if idx == coords.shape[0] - 2 and ratio == 1:
+        return coords, NONE_COORD
+
+    if ratio == 0:
+        seg_0 = coords[:idx + 1] 
+        seg_1 = coords[idx:]    
+    elif ratio < 1:
+        seg_0 = np.concatenate([coords[:idx+1], [proj]]) 
+        seg_1 = np.concatenate([[proj], coords[idx+1:]])     
+    else:
+        seg_0 = coords[:idx+2]
+        seg_1 = coords[idx+1:]
+    
+    return seg_0, seg_1
+
+# @numba.jit
+def numba_cut_lines(col_idxs, closest, ratio, lines):
+    res = [cut_line(i, c, r, s) 
+           for i, c, r, s in zip(col_idxs, closest, ratio, lines)]
+    
+    return res
+
+def linear_referencing_vector(points:np.array, lines:np.array, cut=True, eps=1e-9):
+    n_len = [len(i) for i in lines]
+    n_cols = max(n_len)
+    n_rows = len(lines)
+    _lines, mask = lines_to_matrix(lines, n_rows, n_cols)
+
+    segs = np.dstack([_lines[:, :-1][:,:,np.newaxis],
+                      _lines[:, 1:][:,:,np.newaxis]])
+    pq = segs[:, :, 1] - segs[:, :, 0]
+    d = np.power(pq, 2).sum(axis=-1)
+    len_np = np.sqrt(d)
+    d[d == 0] = eps
+
+    x, y = points[:, 0], points[:, 1]
+    dx = x[:, np.newaxis] - segs[:, :, 0, 0]
+    dy = y[:, np.newaxis] - segs[:, :, 0, 1]
+    t = pq[:, :, 0] * dx + pq[:, :, 1] * dy
+    
+    ratios = t / d
+    ratios[ratios < 0] = 0
+    ratios[ratios > 1] = 1
+
+    offset = pq * ratios[:, :, np.newaxis] # (n, l, 2)
+    closests = offset + segs[:, :, 0]
+    dists = np.linalg.norm(points[:, np.newaxis] - closests, axis=-1)
+
+    col_idxs = np.argmin(dists, axis=1)
+    row_idxs = np.arange(n_rows)
+    
+    cp = closests[row_idxs, col_idxs]
+    r = ratios[row_idxs, col_idxs]
+    dist_p2c = dists[row_idxs, col_idxs]
+
+    sum_mask = np.zeros((n_rows, n_cols-1), dtype=np.bool_)
+    for i, col in enumerate(col_idxs):
+        sum_mask[i, :col] = True
+
+    offset = np.sum(len_np, axis=1, where=sum_mask) + len_np[row_idxs, col_idxs] * r
+
+    res = {}
+    res['proj_point'] = [i for i in cp]
+    res['dist_p2c'] = dist_p2c
+
+    if not cut:
+        res['offset'] = offset
+    else:
+        tmp = numba_cut_lines(col_idxs, cp, r, lines)
+        seg_0, seg_1 = list(zip(*tmp))
+        res['seg_0'] = seg_0
+        res['seg_1'] = seg_1
+        res['len_0'] = offset
+        res['len_1'] = len_np.sum(axis=1) - offset
+
+    return res
+
+def linear_referencing_geom(point_geoms, line_geoms, cut=True, eps=1e-9):
+    _points = points_geoseries_2_ndarray(point_geoms)
+    _lines = geoseries_to_coords(line_geoms)
+
+    res = linear_referencing_vector(_points, _lines, cut, eps)    
 
     return res