+++ mapmatching/utils/timer.py
@@ -36,7 +36,7 @@ def timeit(func):
         res = func(*args, **kwargs)
         end = time.time()
         _log = f"{func.__name__}, cost: {(end - start) * 1000: .2f} ms"
-        print(_log)
+        # print(_log)
         logger.info(_log)
         
         return res